requirements.txt
absl-py==1.0.0
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
ftfy==6.1.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gunicorn==20.0.4


Markdown==3.3.6
MarkupSafe==2.1.1
munkres==1.1.4
nest-asyncio==1.5.4
networkx==2.6.3
newsapi==0.1.1
newsapi-python==0.2.6

oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.0
psycopg2==2.9.3
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.24.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0

